complete program

+++ webDemo/src/main/java/com/dxc/application/feature/common/data/database/model/AttachedFile.java
@@ -2,12 +2,11 @@ package com.dxc.application.feature.common.data.database.model;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
 public class AttachedFile {
-    private BigDecimal id;
+    private Integer id;
     private String fileName;
     private String fileContentType;
     private byte[] fileData;
